move helper function and interface to utils/types
